requirements.txt
aiogram==3.15.0
requests==2.28.1
python-dotenv==1.0.1
cachetools==5.2.0
